# specs/001-agent-core-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Java Shannon Agent Orchestrator Core API
  version: 1.0.0
  description: Shannon 架构对齐的核心接口与事件流定义
servers:
  - url: /
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/mcp/tools/list:
    post:
      summary: MCP 工具列表
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpToolListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/McpToolListResponse'
        '400':
          description: 参数错误，错误码：INVALID_REQUEST、TENANT_MISSING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 访问被拒绝，错误码：POLICY_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在，错误码：NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 服务不可用，错误码：MCP_UNAVAILABLE、CIRCUIT_OPEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/mcp/tools/call:
    post:
      summary: MCP 工具调用
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpToolCallRequest'
      responses:
        '200':
          description: 调用成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/McpToolCallResponse'
        '400':
          description: 参数错误，错误码：INVALID_REQUEST、TENANT_MISSING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 访问被拒绝，错误码：POLICY_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在，错误码：NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Hook 阻断，错误码：HOOK_BLOCKED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 服务不可用，错误码：MCP_UNAVAILABLE、CIRCUIT_OPEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/memory/search:
    post:
      summary: 记忆检索
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryQuery'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/memory/save:
    post:
      summary: 记忆保存
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryRecord'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/memory/compress:
    post:
      summary: 记忆压缩
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompressionRequest'
      responses:
        '200':
          description: 压缩成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/budget/usage:
    post:
      summary: 记录预算使用
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenUsageInput'
      responses:
        '200':
          description: 记录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/budget/summary:
    get:
      summary: 查询预算汇总
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
        traceId:
          type: string
        requestId:
          type: string
    ErrorResponse:
      description: 错误码命名以 POLICY_DENIED、SANDBOX_DENIED 为准，别名 POLICY_DENY、SANDBOX_DENY 仅用于兼容说明。
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        traceId:
          type: string
        requestId:
          type: string
          description: 请求唯一标识（来自请求上下文/请求头生成），用于端到端排查
...
    MemoryRecord:
      type: object
      properties:
        memoryId:
          type: string
        sessionId:
          type: string
        taskId:
          type: string
        content:
          type: string
        summary:
          type: string
        embeddingRef:
          type: string
    MemoryQuery:
      type: object
      properties:
        sessionId:
          type: string
        query:
          type: string
        limit:
          type: integer
...
    CompressionRequest:
      type: object
      properties:
        sessionId:
          type: string
        strategy:
          type: string
...
    TokenUsageInput:
      type: object
      properties:
        usageId:
          type: string
        taskId:
          type: string
        agentId:
          type: string
        model:
          type: string
        provider:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
        costUsd:
          type: number
        tenantId:
          type: string
...
    McpToolDefinition:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        inputSchema:
          type: object
        outputSchema:
          type: object
        tags:
          type: array
          items:
            type: string
    McpToolListRequest:
      type: object
      properties:
        serverId:
          type: string
        cursor:
          type: string
        size:
          type: integer
    McpToolListResponse:
      type: object
      properties:
        tools:
          type: array
          items:
            $ref: '#/components/schemas/McpToolDefinition'
        nextCursor:
          type: string
        hasMore:
          type: boolean
    McpToolCallRequest:
      type: object
      properties:
        callId:
          type: string
        serverId:
          type: string
        toolName:
          type: string
        arguments:
          type: object
        timeoutMs:
          type: integer
    McpToolCallResponse:
      type: object
      properties:
        callId:
          type: string
        status:
          type: string
        result:
          type: object
        error:
          type: object
...